+++ aikido/include/aikido/path/SplineTrajectory2.hpp
@@ -80,7 +80,7 @@ public:
   statespace::StateSpacePtr getStateSpace() const override;
 
   // Documentation inherited.
-  int getNumDerivatives() const override;
+  size_t getNumDerivatives() const override;
 
   // Documentation inherited.
   double getStartTime() const override;
